Benchmark scripts, code refactoring, fixes in parallel alg

+++ lib/temperature_analyzer_serial/include/temperature_analyzer.h
@@ -3,19 +3,9 @@
 
 #include <stdlib.h>
 
-// typedef struct max_delta_t {
-//   int delta;
-//   size_t index;
-
-//   int last_number;
-//   size_t last_index;
-
-//   size_t offset;
-// } max_delta_t;
-
 typedef struct {
   int delta;
-  long long index;
+  long index;
 } delta_temperature_t;
 
 typedef struct {
@@ -24,14 +14,14 @@ typedef struct {
   size_t len;
 
   int prev_elem;
-  long long prev_elem_index;
+  long prev_elem_index;
 
   delta_temperature_t *max_delta_temperature;
 } part_t;
 
 delta_temperature_t init_delta_temp();
-part_t init_part(size_t offset, int *arr, size_t len, int prev_elem,
-                 long long prev_elem_index,
+part_t init_part(long offset, int *arr, size_t len, int prev_elem,
+                 long prev_elem_index,
                  delta_temperature_t *max_delta_temperature);
 
 int find_max_temperature_delta_in_array(part_t *part);
